Ensured latest comments are not overwriten. Refactored OrganisationRegistrationSubmissionService to take the Synchronisation code out of the main file
